Remove DynamoDBCompactionStatusStoreResources
