+++ java/cdk/src/main/java/sleeper/cdk/stack/compaction/CompactionStatusStoreResources.java
@@ -33,15 +33,7 @@ public interface CompactionStatusStoreResources {
 
     static CompactionStatusStoreResources from(Construct scope, String id, InstanceProperties properties, ManagedPoliciesStack policiesStack) {
         if (properties.getBoolean(COMPACTION_STATUS_STORE_ENABLED)) {
-            return new CompactionStatusStoreStack(scope, id, properties, policiesStack).getResources();
-        } else {
-            return none();
-        }
-    }
-
-    static CompactionStatusStoreResources from(Construct scope, InstanceProperties properties, ManagedPoliciesStack policiesStack) {
-        if (properties.getBoolean(COMPACTION_STATUS_STORE_ENABLED)) {
-            return new DynamoDBCompactionStatusStoreResources(scope, properties, policiesStack);
+            return new CompactionStatusStoreStack(scope, id, properties, policiesStack);
         } else {
             return none();
         }
